moved helper funcitons to new file
